change_new

+++ controllers/ActivityController.php
@@ -458,7 +458,7 @@ class ActivityController extends BackendController
         $model = new Activity;
 
         if ($model->load($data, '') && $model->save()) {
-            if (!empty($data['space_spot_id'])) {
+            if (!empty($data['space_spot_id']) && isset($data['space_section_id'])) {
                 if ($data['space_section_id'] > 0) {
                     foreach ($data['space_section_id'] as $space_section) {
                         $r_activity_space =new RActivitySpace();
@@ -841,8 +841,9 @@ class ActivityController extends BackendController
         }
 
         if ($model->save()) {
+            // print_r(isset($data['space_section_id']));
             // 当场地id不为空时
-            if (!empty($data['space_spot_id'])) {
+            if (!empty($data['space_spot_id']) && isset($data['space_section_id']) ) {
                 // 当空间没有选择时默认存储全部
                 if ($data['space_section_id'] > 0) {
                     $delete_spaces = RActivitySpace::deleteAll([
